% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pce.R
\name{boot}
\alias{boot}
\title{Bootstrap variance estimator}
\usage{
boot(
  Z,
  X,
  S,
  Y,
  n_boot = 500,
  n_divisions = 100,
  copula_type = "gaussian",
  rho = 0,
  weighting_function_vectorized = weighting_function_vectorized,
  g_function_vectorized = g_function_vectorized
)
}
\arguments{
\item{Z}{to be added}
}
\value{
estimated_result to be added
}
\description{
Bootstrap variance estimator
}
\examples{
boot()
}
