% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pce.R
\name{matrix_multiply_with_expansion}
\alias{matrix_multiply_with_expansion}
\title{Two matrix multiplication with different first dimension}
\usage{
matrix_multiply_with_expansion(U, V)
}
\arguments{
\item{U}{A matrix of size (n_u, n)}

\item{V}{A matrix of size (n_v, n)}
}
\value{
A matrix of size (n_u, n_v, n)
}
\description{
Two matrix multiplication with different first dimension
}
