% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pce.R
\name{point_estimator}
\alias{point_estimator}
\title{Point estimator}
\usage{
point_estimator(
  Z,
  X,
  S,
  Y,
  n_divisions = 100,
  copula_function_vectorized = copula_function_vectorized,
  copula_gradient_vectorized = copula_gradient_vectorized,
  weighting_function_vectorized = weighting_function_vectorized,
  g_function_vectorized = g_function_vectorized
)
}
\arguments{
\item{Z}{to be added}
}
\value{
estimated_result to be added
}
\description{
Point estimator
}
\examples{
point_estimator()
}
