% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pce.R
\name{trapz_cdf}
\alias{trapz_cdf}
\title{Compute cdf in vector form}
\usage{
trapz_cdf(x, y, axis = 1)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector or a matrix}
}
\value{
if y is a vector, return a number else return a vector
}
\description{
Compute cdf in vector form
}
