% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pce.R
\name{trapz_vector}
\alias{trapz_vector}
\title{Compute integral in vector form}
\usage{
trapz_vector(x, y, axis = 1)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector or a matrix}
}
\value{
if y is a vector, return a number else return a vector
}
\description{
Compute integral in vector form
}
